requirements-prod.txt
# ============================================================================
# Production Requirements - Khalifa Pharmacy System
# متطلبات الإنتاج - نظام صيدليات خليفة
# ============================================================================
# Created by: محمد فارس
# ============================================================================

# Include all base requirements
-r requirements.txt

# ============================================================================
# Production Web Server
# ============================================================================
gunicorn==21.2.0
uvicorn==0.25.0

# ============================================================================
# Database Drivers (Production)
# ============================================================================
# PostgreSQL (Recommended)
psycopg2-binary==2.9.9

# MySQL (Alternative)
# mysqlclient==2.2.0

# ============================================================================
# Caching & Queue Management
# ============================================================================
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# ============================================================================
# WebSocket Support (Real-time)
# ============================================================================
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# ============================================================================
# Security & Monitoring
# ============================================================================
sentry-sdk==1.39.2
django-ratelimit==4.1.0

# ============================================================================
# Static Files & Media Storage
# ============================================================================
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.10  # For AWS S3

# ============================================================================
# Email Backend
# ============================================================================
django-anymail==10.2

# ============================================================================
# Performance & Optimization
# ============================================================================
django-cachalot==2.6.1

# ============================================================================
# Health Checks
# ============================================================================
django-health-check==3.18.1

# ============================================================================
# Environment Variables
# ============================================================================
python-decouple==3.8

# ============================================================================
# Notes:
# ============================================================================
# Install production requirements:
#   pip install -r requirements-prod.txt
#
# Run with Gunicorn:
#   gunicorn khalifa_pharmacy.wsgi:application --bind 0.0.0.0:8000 --workers 4
#
# Run Celery Worker:
#   celery -A khalifa_pharmacy worker -l info
#
# Run Celery Beat (Scheduler):
#   celery -A khalifa_pharmacy beat -l info
#
# Run with Daphne (WebSocket):
#   daphne -b 0.0.0.0 -p 8000 khalifa_pharmacy.asgi:application
# ============================================================================
